Cybersecurity-Log-Analyser/requirements.txt
Flask==2.3.4
twilio==8.16.0
pandas==2.1.1
matplotlib==3.8.0
sqlite3
python-dotenv==1.0.0
